requirement.txt
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
ffmpeg-python==0.2.0
filelock==3.7.1
future==0.18.2
huggingface-hub==0.8.1
idna==3.3
numpy==1.23.0
packaging==21.3
pycparser==2.21
pydub==0.25.1
pyparsing==3.0.9
PyYAML==6.0
regex==2022.7.9
requests==2.28.1
sounddevice==0.4.4
tokenizers==0.12.1
torch==1.12.0
tqdm==4.64.0
transformers==4.20.1
typing_extensions==4.3.0
urllib3==1.26.10
vosk==0.3.38
